# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/accounts/{accountId}/transactions": {
      "post": {
        "operationId": "TransactionController_create",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "amount": {
                    "maxLength": 1002,
                    "pattern": "^\\d+(?:\\.\\d+)?$",
                    "type": "string"
                  },
                  "currencyCode": {
                    "pattern": "^[A-Z]{3}$",
                    "type": "string"
                  },
                  "currencyMinorUnitScale": {
                    "maximum": 18,
                    "minimum": 0,
                    "type": "integer"
                  },
                  "description": {
                    "nullable": true,
                    "type": "string"
                  },
                  "kind": {
                    "enum": ["income", "expense"],
                    "type": "string"
                  },
                  "observations": {
                    "nullable": true,
                    "type": "string"
                  },
                  "occurredAt": {
                    "format": "date-time",
                    "type": "string"
                  }
                },
                "required": [
                  "amount",
                  "currencyCode",
                  "currencyMinorUnitScale",
                  "description",
                  "kind",
                  "observations",
                  "occurredAt"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "accountId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "amount": {
                      "pattern": "^\\d+(?:\\.\\d+)?$",
                      "type": "string"
                    },
                    "archivedAt": {
                      "format": "date-time",
                      "nullable": true,
                      "type": "string"
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "currencyCode": {
                      "pattern": "^[A-Z]{3}$",
                      "type": "string"
                    },
                    "currencyMinorUnitScale": {
                      "maximum": 18,
                      "minimum": 0,
                      "type": "integer"
                    },
                    "description": {
                      "nullable": true,
                      "type": "string"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "kind": {
                      "enum": ["income", "expense"],
                      "type": "string"
                    },
                    "lifecycle": {
                      "enum": ["active", "archived", "trashed"],
                      "type": "string"
                    },
                    "observations": {
                      "nullable": true,
                      "type": "string"
                    },
                    "occurredAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "trashedAt": {
                      "format": "date-time",
                      "nullable": true,
                      "type": "string"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "version": {
                      "minimum": 1,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "accountId",
                    "amount",
                    "archivedAt",
                    "createdAt",
                    "currencyCode",
                    "currencyMinorUnitScale",
                    "description",
                    "id",
                    "kind",
                    "lifecycle",
                    "observations",
                    "occurredAt",
                    "trashedAt",
                    "updatedAt",
                    "version"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed before recording a transaction",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "enum": ["INVALID_REQUEST"],
                          "type": "string"
                        },
                        "correlationId": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "correlationId", "message"],
                      "type": "object"
                    }
                  },
                  "required": ["error"],
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing or invalid"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record an income or expense declared by the authenticated user",
        "tags": ["transactions"]
      },
      "get": {
        "operationId": "TransactionController_list",
        "parameters": [
          {
            "name": "lifecycle",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["active", "archived", "trashed"],
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "accountId": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "amount": {
                        "pattern": "^\\d+(?:\\.\\d+)?$",
                        "type": "string"
                      },
                      "archivedAt": {
                        "format": "date-time",
                        "nullable": true,
                        "type": "string"
                      },
                      "createdAt": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "currencyCode": {
                        "pattern": "^[A-Z]{3}$",
                        "type": "string"
                      },
                      "currencyMinorUnitScale": {
                        "maximum": 18,
                        "minimum": 0,
                        "type": "integer"
                      },
                      "description": {
                        "nullable": true,
                        "type": "string"
                      },
                      "id": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "kind": {
                        "enum": ["income", "expense"],
                        "type": "string"
                      },
                      "lifecycle": {
                        "enum": ["active", "archived", "trashed"],
                        "type": "string"
                      },
                      "observations": {
                        "nullable": true,
                        "type": "string"
                      },
                      "occurredAt": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "trashedAt": {
                        "format": "date-time",
                        "nullable": true,
                        "type": "string"
                      },
                      "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "version": {
                        "minimum": 1,
                        "type": "integer"
                      }
                    },
                    "required": [
                      "accountId",
                      "amount",
                      "archivedAt",
                      "createdAt",
                      "currencyCode",
                      "currencyMinorUnitScale",
                      "description",
                      "id",
                      "kind",
                      "lifecycle",
                      "observations",
                      "occurredAt",
                      "trashedAt",
                      "updatedAt",
                      "version"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing or invalid"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List transactions for an owned account",
        "tags": ["transactions"]
      }
    },
    "/api/v1/transactions": {
      "get": {
        "operationId": "TransactionController_listBetween",
        "parameters": [
          {
            "name": "occurredAtOrder",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["asc", "desc"],
              "type": "string"
            }
          },
          {
            "name": "lifecycle",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["active", "archived", "trashed"],
              "type": "string"
            }
          },
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["income", "expense"],
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "accountId": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "amount": {
                        "pattern": "^\\d+(?:\\.\\d+)?$",
                        "type": "string"
                      },
                      "archivedAt": {
                        "format": "date-time",
                        "nullable": true,
                        "type": "string"
                      },
                      "createdAt": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "currencyCode": {
                        "pattern": "^[A-Z]{3}$",
                        "type": "string"
                      },
                      "currencyMinorUnitScale": {
                        "maximum": 18,
                        "minimum": 0,
                        "type": "integer"
                      },
                      "description": {
                        "nullable": true,
                        "type": "string"
                      },
                      "id": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "kind": {
                        "enum": ["income", "expense"],
                        "type": "string"
                      },
                      "lifecycle": {
                        "enum": ["active", "archived", "trashed"],
                        "type": "string"
                      },
                      "observations": {
                        "nullable": true,
                        "type": "string"
                      },
                      "occurredAt": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "trashedAt": {
                        "format": "date-time",
                        "nullable": true,
                        "type": "string"
                      },
                      "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "version": {
                        "minimum": 1,
                        "type": "integer"
                      }
                    },
                    "required": [
                      "accountId",
                      "amount",
                      "archivedAt",
                      "createdAt",
                      "currencyCode",
                      "currencyMinorUnitScale",
                      "description",
                      "id",
                      "kind",
                      "lifecycle",
                      "observations",
                      "occurredAt",
                      "trashedAt",
                      "updatedAt",
                      "version"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing or invalid"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List owned transactions in an explicit instant range",
        "tags": ["transactions"]
      }
    },
...
  },
  "info": {
    "title": "Seshat Finance API",
    "description": "API para registrar e organizar informações financeiras.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}